repeatRoutine 메서드 분리 및 step1, step2 파일 생성
